Moved setup stuff from separate tests to setUp in compaction_test
